+++ rllab/distributions/diagonal_gaussian.py
@@ -4,6 +4,12 @@ from rllab.distributions.base import Distribution
 
 
 class DiagonalGaussian(Distribution):
+    def __init__(self, dim):
+        self._dim = dim
+
+    @property
+    def dim(self):
+        return self._dim
 
     def kl_sym(self, old_dist_info_vars, new_dist_info_vars):
         old_means = old_dist_info_vars["mean"]